requirements.txt
feedparser==5.2.1
pandas==0.25.2
google-cloud-bigquery==1.22.0